+++ src/main/java/com/siddhiApi/controller/StreamsController.java
@@ -1,6 +1,7 @@
 package com.siddhiApi.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.siddhiApi.authorization.ApiKeyAuth;
 import com.siddhiApi.entity.Stream;
 import com.siddhiApi.entity.Subscription;
 import com.siddhiApi.exceptions.DuplicatedEntity;
@@ -29,9 +30,14 @@ public class StreamsController {
     @Autowired
     private StreamService streamService;
 
+    private final ApiKeyAuth apiKeyAuth = ApiKeyAuth.getApiKeyAuth();
+
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping()
-    public void createStream(@RequestBody Stream stream){
+    public void createStream(@RequestBody Stream stream, @RequestHeader(required = false, name = "X-API-Key") String api_key){
+        if (!apiKeyAuth.auth(api_key)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "The api key introduced is not valid");
+        }
         try {
             streamService.createStream(stream);
         } catch (DuplicatedEntity duplicatedEntity) {
@@ -40,14 +46,20 @@ public class StreamsController {
     }
 
     @GetMapping(value="", produces = MediaType.APPLICATION_JSON_VALUE)
-    public String[] getStream(){
+    public String[] getStream(@RequestHeader(required = false, name = "X-API-Key") String api_key){
+        if (!apiKeyAuth.auth(api_key)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "The api key introduced is not valid");
+        }
         return Arrays.stream(streamService.getStreams())
                 .map(Stream::getStreamID)
                 .toArray(String[]::new);
     }
 
     @GetMapping(value = "/{name}", produces = MediaType.APPLICATION_JSON_VALUE)
-    public String getStream(@PathVariable String name){
+    public String getStream(@PathVariable String name, @RequestHeader(required = false, name = "X-API-Key") String api_key){
+        if (!apiKeyAuth.auth(api_key)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "The api key introduced is not valid");
+        }
         try {
             return streamService.getStream(name).getJsonSchema().toString();
         } catch (NotFoundException e) {
@@ -57,7 +69,10 @@ public class StreamsController {
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping(value = "/{name}")
-    public void removeStream(@PathVariable String name){
+    public void removeStream(@PathVariable String name, @RequestHeader(required = false, name = "X-API-Key") String api_key){
+        if (!apiKeyAuth.auth(api_key)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "The api key introduced is not valid");
+        }
         try {
             streamService.removeStream(name);
         } catch (NotFoundException e) {
@@ -69,8 +84,11 @@ public class StreamsController {
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping(value = "/{streamID}/events")
-    public void sendEvent(@PathVariable String streamID, @RequestBody Object event){
+    public void sendEvent(@PathVariable String streamID, @RequestBody Object event, @RequestHeader(required = false, name = "X-API-Key") String api_key){
         logger.info("Printing event schema: " + event);
+        if (!apiKeyAuth.auth(api_key)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "The api key introduced is not valid");
+        }
         try {
             streamService.sendEvent(streamID, event);
         } catch (JsonProcessingException e) {
@@ -84,7 +102,10 @@ public class StreamsController {
 
     @ResponseStatus(HttpStatus.CREATED)
     @PostMapping(value = "/{streamID}/subscriptions")
-    public Subscription subscribe(@PathVariable String streamID, @RequestBody Subscription subscription){
+    public Subscription subscribe(@PathVariable String streamID, @RequestBody Subscription subscription, @RequestHeader(required = false, name = "X-API-Key") String api_key){
+        if (!apiKeyAuth.auth(api_key)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "The api key introduced is not valid");
+        }
         try {
             return streamService.subscribe(streamID, subscription);
         } catch (NotFoundException e) {
@@ -93,7 +114,10 @@ public class StreamsController {
     }
 
     @GetMapping(value = "/{streamID}/subscriptions")
-    public Subscription[] getStreamSubscriptions(@PathVariable String streamID) {
+    public Subscription[] getStreamSubscriptions(@PathVariable String streamID, @RequestHeader(required = false, name = "X-API-Key") String api_key) {
+        if (!apiKeyAuth.auth(api_key)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "The api key introduced is not valid");
+        }
         try {
             return streamService.getSubscriptions(streamID);
         } catch (NotFoundException notFoundException) {
@@ -102,7 +126,10 @@ public class StreamsController {
     }
 
     @GetMapping(value = "/{streamID}/subscriptions/{id}")
-    public Subscription getStreamSubscription(@PathVariable String streamID, @PathVariable String id){
+    public Subscription getStreamSubscription(@PathVariable String streamID, @PathVariable String id, @RequestHeader(required = false, name = "X-API-Key") String api_key){
+        if (!apiKeyAuth.auth(api_key)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "The api key introduced is not valid");
+        }
         try {
             return streamService.getSubscription(streamID, id);
         } catch (NotFoundException notFoundException) {
@@ -112,7 +139,10 @@ public class StreamsController {
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @DeleteMapping(value = "/{streamID}/subscriptions/{id}")
-    public void deleteStreamSubscription(@PathVariable String streamID, @PathVariable String id){
+    public void deleteStreamSubscription(@PathVariable String streamID, @PathVariable String id, @RequestHeader(required = false, name = "X-API-Key") String api_key){
+        if (!apiKeyAuth.auth(api_key)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "The api key introduced is not valid");
+        }
         try {
             streamService.unsubscribe(streamID, id);
         } catch (NotFoundException notFoundException) {
@@ -122,7 +152,11 @@ public class StreamsController {
 
     @ResponseStatus(HttpStatus.NO_CONTENT)
     @PutMapping(value = "/{streamID}/subscriptions/{subscriptionID}")
-    public void updateStreamSubscription(@PathVariable String streamID, @PathVariable String subscriptionID, @RequestBody Subscription subscription){
+    public void updateStreamSubscription(@PathVariable String streamID, @PathVariable String subscriptionID,
+                                         @RequestBody Subscription subscription, @RequestHeader(required = false, name = "X-API-Key") String api_key){
+        if (!apiKeyAuth.auth(api_key)){
+            throw new ResponseStatusException(HttpStatus.FORBIDDEN, "The api key introduced is not valid");
+        }
         try {
             streamService.updateSubscription(streamID, subscriptionID, subscription);
         } catch (NotFoundException notFoundException) {
